Move root around a bit
